# [EC] Unify point addition for P-256/384/521